requirements.txt
Django==3.0.3
djangorestframework==3.11.0
Markdown==3.3.3
django-filter==2.4.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.17
